Slider: Re-enable resize listener
We don't need that whole state update on resize stuff anymore because
all of those values are now being calculated by the render method. But
we do need to tell the component to re-render on window resize, otherwise
the translateValue would not get updated until the next state update (e.g.
page change).